refactor : 소스 리팩토링
