+++ src/main/java/me/deftkang/springbootdeveloper/domain/Article.java
@@ -12,7 +12,9 @@ import org.hibernate.annotations.Type;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 
+import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.time.temporal.ChronoUnit;
 import java.util.UUID;
 
 @Entity
@@ -42,5 +44,14 @@ public class Article extends BaseEntity {
     public void update(String title, String content) {
         this.title = title;
         this.content = content;
+    }
+
+    public long calculateModifiableDate(LocalDateTime createdDatedAt) {
+        LocalDate localDate = LocalDate.now();
+        long diffDate = ChronoUnit.DAYS.between(createdDatedAt.toLocalDate(), localDate);
+        //생성일 - 현재일이 9일이면 하루 남은걸로 한다.
+        long modifiableDate = (10 - diffDate) > 0 ? (10 - diffDate) : 0;
+
+        return modifiableDate;
     }
 }
